personalized_shopping_agent/pyproject.toml
[project]
name = "personalized-shopping"
version = "1.0.0"
description = "A2A-compatible Personalized Shopping Agent (simplified demo)"
authors = [{ name = "A2A Test", email = "test@example.com" }]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk[a2a]>=1.0.0",
    "litellm>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["personalized_shopping"]
